const correctness & coding style fixes, tora tha arxiso na kano kai tipota xrisimo :)
--HG--
branch : ubertk
extra : convert_revision : svn%3A419d067c-91e1-11de-9a8c-5f207592b8ba/ubertk%407

+++ src/utk_container.cc
@@ -165,13 +165,13 @@ void Container::draw() const
 
 int Container::get_width() const
 {
-	unsigned int w = 0;
-	for (unsigned int i=0; i<cont.size(); i++)
+	size_t w = 0;
+	for (size_t i=0; i<cont.size(); i++)
 	{
 		w += (*this)[i]->get_width() + spacing;
 	}
 	w += 2 * border;
-	return w;
+	return (int)w;
 }
 
 // TODO: also change the size of the container to fit the children
